Added methods
NOTE: nuke.py and config.py are one tool. config.py is used to compile what you want to spam and etc. edit config.py to your liking.
NOTE2: the batch file is not a scam, and UntitledNuker is not my tool.
NOTE3: REKTYHOOK.py is very effective, have fun with them all!